DOP-1534: Add more granular testing framework
